+++ share-my-event-api/src/main/java/co/simplon/p25/sharemyeventapi/services/AuthServiceImpl.java
@@ -18,7 +18,7 @@ public class AuthServiceImpl implements AuthService {
 	}
 
 	@Override
-	public int findActorIdByAuthId() {
+	public Long findActorIdByAuthId() {
 		return actorRepository
 				.findActorIdByAuthId(SecurityHelper.authenticatedUserId());
 	}
